Collapsing features speedup using numpy implementation. The net runtime for collapsing the features in mimic (964251 rows and 15 features) goes from ~4 hours to ~5 minutes. 

To run the script : 
- Go to src/logistic/
- Run the script full_pipeline.sh : ./full_pipeline.sh
- The script runs the feature_transformation.py python script. The runtime for each collapse function for the different percentile ranges are also displayed.

```
>> (tspred_env) [prath01@omega001 logistic]$ ./full_pipeline.sh
Align to grid
Normalize Features
Feature transformations collapse
done parsing...
reading csv...
done reading csv...
collapsing data..
collapsing with func count in 0 to 10 percentile range
time taken : 0 seconds
collapsing with func count in 0 to 25 percentile range
time taken : 0 seconds
collapsing with func count in 0 to 50 percentile range
time taken : 0 seconds
collapsing with func count in 50 to 100 percentile range
time taken : 0 seconds
collapsing with func count in 75 to 100 percentile range
time taken : 0 seconds
collapsing with func count in 90 to 100 percentile range
time taken : 0 seconds
collapsing with func count in 0 to 100 percentile range
time taken : 0 seconds
collapsing with func mean in 0 to 10 percentile range
time taken : 2 seconds
collapsing with func mean in 0 to 25 percentile range
time taken : 2 seconds
collapsing with func mean in 0 to 50 percentile range
time taken : 2 seconds
collapsing with func mean in 50 to 100 percentile range
time taken : 2 seconds
collapsing with func mean in 75 to 100 percentile range
time taken : 2 seconds
collapsing with func mean in 90 to 100 percentile range
time taken : 2 seconds
collapsing with func mean in 0 to 100 percentile range
time taken : 2 seconds
collapsing with func median in 0 to 10 percentile range
time taken : 12 seconds
collapsing with func median in 0 to 25 percentile range
time taken : 12 seconds
collapsing with func median in 0 to 50 percentile range
time taken : 12 seconds
collapsing with func median in 50 to 100 percentile range
time taken : 12 seconds
collapsing with func median in 75 to 100 percentile range
time taken : 12 seconds
collapsing with func median in 90 to 100 percentile range
time taken : 12 seconds
collapsing with func median in 0 to 100 percentile range
time taken : 13 seconds
collapsing with func std in 0 to 10 percentile range
time taken : 3 seconds
collapsing with func std in 0 to 25 percentile range
time taken : 3 seconds
collapsing with func std in 0 to 50 percentile range
time taken : 3 seconds
collapsing with func std in 50 to 100 percentile range
time taken : 3 seconds
collapsing with func std in 75 to 100 percentile range
time taken : 3 seconds
collapsing with func std in 90 to 100 percentile range
time taken : 3 seconds
collapsing with func std in 0 to 100 percentile range
time taken : 3 seconds
collapsing with func min in 0 to 10 percentile range
time taken : 1 seconds
collapsing with func min in 0 to 25 percentile range
time taken : 1 seconds
collapsing with func min in 0 to 50 percentile range
time taken : 1 seconds
collapsing with func min in 50 to 100 percentile range
time taken : 1 seconds
collapsing with func min in 75 to 100 percentile range
time taken : 1 seconds
collapsing with func min in 90 to 100 percentile range
time taken : 1 seconds
collapsing with func min in 0 to 100 percentile range
time taken : 1 seconds
collapsing with func max in 0 to 10 percentile range
time taken : 1 seconds
collapsing with func max in 0 to 25 percentile range
time taken : 1 seconds
collapsing with func max in 0 to 50 percentile range
time taken : 1 seconds
collapsing with func max in 50 to 100 percentile range
time taken : 1 seconds
collapsing with func max in 75 to 100 percentile range
time taken : 1 seconds
collapsing with func max in 90 to 100 percentile range
time taken : 1 seconds
collapsing with func max in 0 to 100 percentile range
time taken : 1 seconds
collapsing with func slope in 0 to 10 percentile range
time taken : 23 seconds
collapsing with func slope in 0 to 25 percentile range
time taken : 24 seconds
collapsing with func slope in 0 to 50 percentile range
time taken : 24 seconds
collapsing with func slope in 50 to 100 percentile range
time taken : 24 seconds
collapsing with func slope in 75 to 100 percentile range
time taken : 23 seconds
collapsing with func slope in 90 to 100 percentile range
time taken : 23 seconds
collapsing with func slope in 0 to 100 percentile range
time taken : 24 seconds
-----------------------------------------
total time taken = 319 seconds
-----------------------------------------
done collapsing data..
time taken to collapse data : 321.0333790779114 seconds
saving data to output...
```